cortexdb-py/pyproject.toml
[tool.poetry]
name = "cortexdb"
version = "0.1.0"
description = "Python SDK for CortexDB — a RAG-powered memory database"
authors = ["VectorNode"]
readme = "README.md"
packages = [{include = "cortexdb"}]

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.28.0"
pydantic = "^2.0"
google-genai = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.25.0"
respx = "^0.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
